fluss-common/src/test/java/com/alibaba/fluss/cluster/ClusterTest.java
/*
 * Copyright (c) 2024 Alibaba Group Holding Ltd.
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *      http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

package com.alibaba.fluss.cluster;

import com.alibaba.fluss.config.Configuration;
import com.alibaba.fluss.metadata.PhysicalTablePath;
import com.alibaba.fluss.metadata.SchemaInfo;
import com.alibaba.fluss.metadata.TableInfo;
import com.alibaba.fluss.metadata.TablePath;

import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;

import java.util.Arrays;
import java.util.Collections;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

import static com.alibaba.fluss.record.TestData.DATA1_PHYSICAL_TABLE_PATH;
import static com.alibaba.fluss.record.TestData.DATA1_SCHEMA;
import static com.alibaba.fluss.record.TestData.DATA1_TABLE_ID;
import static com.alibaba.fluss.record.TestData.DATA1_TABLE_INFO;
import static com.alibaba.fluss.record.TestData.DATA1_TABLE_PATH;
import static com.alibaba.fluss.record.TestData.DATA2_SCHEMA;
import static com.alibaba.fluss.record.TestData.DATA2_TABLE_ID;
import static com.alibaba.fluss.record.TestData.DATA2_TABLE_INFO;
import static com.alibaba.fluss.record.TestData.DATA2_TABLE_PATH;
import static org.assertj.core.api.Assertions.assertThat;
import static org.assertj.core.api.Assertions.assertThatThrownBy;

/** Test for {@link Cluster}. */
class ClusterTest {
    private static final ServerNode COORDINATOR_SERVER =
            new ServerNode(-1, "localhost", 98, ServerType.COORDINATOR);
    private static final ServerNode[] NODES =
            new ServerNode[] {
                new ServerNode(0, "localhost", 99, ServerType.TABLET_SERVER),
                new ServerNode(1, "localhost", 100, ServerType.TABLET_SERVER),
                new ServerNode(2, "localhost", 101, ServerType.TABLET_SERVER),
                new ServerNode(11, "localhost", 102, ServerType.TABLET_SERVER)
            };

    private Configuration conf;
    private Map<Integer, ServerNode> aliveTabletServersById;

    @BeforeEach
    void setup() {
        aliveTabletServersById = new HashMap<>();
        for (ServerNode node : NODES) {
            aliveTabletServersById.put(node.id(), node);
        }
    }

    @Test
    void testReturnModifiableCollections() {
        Map<PhysicalTablePath, List<BucketLocation>> tablePathToBucketLocations = new HashMap<>();
        tablePathToBucketLocations.put(
                DATA1_PHYSICAL_TABLE_PATH,
                Arrays.asList(
                        new BucketLocation(
                                DATA1_PHYSICAL_TABLE_PATH, DATA1_TABLE_ID, 0, NODES[0], NODES),
                        new BucketLocation(
                                DATA1_PHYSICAL_TABLE_PATH, DATA1_TABLE_ID, 1, null, NODES),
                        new BucketLocation(
                                DATA1_PHYSICAL_TABLE_PATH, DATA1_TABLE_ID, 2, NODES[2], NODES)));
        tablePathToBucketLocations.put(
                PhysicalTablePath.of(DATA2_TABLE_PATH),
                Arrays.asList(
                        new BucketLocation(
                                PhysicalTablePath.of(DATA2_TABLE_PATH),
                                DATA2_TABLE_ID,
                                0,
                                null,
                                NODES),
                        new BucketLocation(
                                PhysicalTablePath.of(DATA2_TABLE_PATH),
                                DATA2_TABLE_ID,
                                1,
                                NODES[0],
                                NODES)));

        Map<TablePath, Long> tablePathToTableId = new HashMap<>();
        tablePathToTableId.put(DATA1_TABLE_PATH, DATA1_TABLE_ID);
        tablePathToTableId.put(DATA2_TABLE_PATH, DATA2_TABLE_ID);

        Map<TablePath, TableInfo> tablePathToTableInfo = new HashMap<>();
        tablePathToTableInfo.put(DATA1_TABLE_PATH, DATA1_TABLE_INFO);
        tablePathToTableInfo.put(DATA2_TABLE_PATH, DATA2_TABLE_INFO);

        Cluster cluster =
                new Cluster(
                        aliveTabletServersById,
                        COORDINATOR_SERVER,
                        tablePathToBucketLocations,
                        tablePathToTableId,
                        Collections.emptyMap(),
                        tablePathToTableInfo);

        assertThatThrownBy(() -> cluster.getAliveTabletServers().put(1, NODES[3]))
                .isInstanceOf(UnsupportedOperationException.class);
        assertThatThrownBy(
                        () ->
                                cluster.getAvailableBucketsForPhysicalTablePath(
                                                DATA1_PHYSICAL_TABLE_PATH)
                                        .add(
                                                new BucketLocation(
                                                        DATA1_PHYSICAL_TABLE_PATH,
                                                        DATA1_TABLE_ID,
                                                        3,
                                                        NODES[3],
                                                        NODES)))
                .isInstanceOf(UnsupportedOperationException.class);
    }

    @Test
    void testGetTable() {
        Map<PhysicalTablePath, List<BucketLocation>> tablePathToBucketLocations = new HashMap<>();
        tablePathToBucketLocations.put(
                DATA1_PHYSICAL_TABLE_PATH,
                Arrays.asList(
                        new BucketLocation(
                                DATA1_PHYSICAL_TABLE_PATH, DATA1_TABLE_ID, 0, NODES[0], NODES),
                        new BucketLocation(
                                DATA1_PHYSICAL_TABLE_PATH, DATA1_TABLE_ID, 1, null, NODES),
                        new BucketLocation(
                                DATA1_PHYSICAL_TABLE_PATH, DATA1_TABLE_ID, 2, NODES[2], NODES)));
        tablePathToBucketLocations.put(
                PhysicalTablePath.of(DATA2_TABLE_PATH),
                Arrays.asList(
                        new BucketLocation(
                                PhysicalTablePath.of(DATA2_TABLE_PATH),
                                DATA2_TABLE_ID,
                                0,
                                null,
                                NODES),
                        new BucketLocation(
                                PhysicalTablePath.of(DATA2_TABLE_PATH),
                                DATA2_TABLE_ID,
                                1,
                                NODES[0],
                                NODES)));

        Map<TablePath, Long> tablePathToTableId = new HashMap<>();
        tablePathToTableId.put(DATA1_TABLE_PATH, DATA1_TABLE_ID);
        tablePathToTableId.put(DATA2_TABLE_PATH, DATA2_TABLE_ID);

        Map<TablePath, TableInfo> tablePathToTableInfo = new HashMap<>();
        tablePathToTableInfo.put(DATA1_TABLE_PATH, DATA1_TABLE_INFO);
        tablePathToTableInfo.put(DATA2_TABLE_PATH, DATA2_TABLE_INFO);

        Cluster cluster =
                new Cluster(
                        aliveTabletServersById,
                        COORDINATOR_SERVER,
                        tablePathToBucketLocations,
                        tablePathToTableId,
                        Collections.emptyMap(),
                        tablePathToTableInfo);

        assertThat(cluster.getTable(DATA1_TABLE_PATH).get()).isEqualTo(DATA1_TABLE_INFO);
        assertThat(cluster.getTable(DATA2_TABLE_PATH).get()).isEqualTo(DATA2_TABLE_INFO);
        assertThat(cluster.getSchema(DATA1_TABLE_PATH).get())
                .isEqualTo(new SchemaInfo(DATA1_SCHEMA, 1));
        assertThat(cluster.getSchema(DATA2_TABLE_PATH).get())
                .isEqualTo(new SchemaInfo(DATA2_SCHEMA, 1));
    }
}
